separate wled handlers into classes
