+++ handlers/main_http_handler.py
@@ -7,11 +7,11 @@ from time import sleep
 from aiohttp import web
 
 from handlers.spotify_api_handler import SpotifyAPIHandler
+from handlers.wled import WLEDArtNet, WLEDJson
 from utils.common import WLEDMode
 from utils.image_utils import downscale_image, image_to_rgb_array, download_image
 from utils.spotify_utils import calculate_remaining_time
 
-from handlers.wled_handler import WLEDArtNet, WLEDJson
 
 # Spotify API has a rate limit per 30-second rolling window
 # so we have to be conservative in the polling frequency
