+++ handlers/wled/__init__.py
@@ -1,0 +1,5 @@
+"""
+classes for WLED handler
+"""
+from handlers.wled.wled_json import WLEDJson as WLEDJson
+from handlers.wled.wled_artnet import WLEDArtNet as WLEDArtNet
